Fix/env var (#6)
* added readme.md and config to be available from environmental variables

* added config doc

* added config to take envvars for sensitive data, fixed directory name typo
