+++ src/lib/utility.js
@@ -37,7 +37,7 @@
 
 const Config = require('../lib/config')
 const Mustache = require('mustache')
-const KafkaConfig = Config.KAFKA
+const KafkaConfig = require('../../config/default.json').KAFKA
 const Logger = require('@mojaloop/central-services-shared').Logger
 const Uuid = require('uuid4')
 const Kafka = require('./kafka')
